+++ lib/CalDAV/Schedule/IMipPlugin.php
@@ -30,13 +30,6 @@ class IMipPlugin extends DAV\ServerPlugin {
     protected $senderEmail;
 
     /**
-     * Server class
-     *
-     * @var DAV\Server
-     */
-    protected $server;
-
-    /**
      * ITipMessage
      *
      * @var ITip\Message
@@ -158,7 +151,7 @@ class IMipPlugin extends DAV\ServerPlugin {
     // This is deemed untestable in a reasonable manner
 
     /**
-     * This function is reponsible for sending the actual email.
+     * This function is responsible for sending the actual email.
      *
      * @param string $to Recipient email address
      * @param string $subject Subject of the email