+++ lib/CalDAV/Schedule/Plugin.php
@@ -529,7 +529,7 @@ class Plugin extends ServerPlugin {
 
             if (isset($newObject->VEVENT->ORGANIZER) && ($newObject->VEVENT->ORGANIZER->getNormalizedValue() === $message->recipient)) {
                 if ($message->scheduleStatus) {
-                    $newObject->VEVENT->ORGANIZER['SCHEDULE-STATUS'] = $message->scheduleStatus;
+                    $newObject->VEVENT->ORGANIZER['SCHEDULE-STATUS'] = $message->getScheduleStatus();
                 }
                 unset($newObject->VEVENT->ORGANIZER['SCHEDULE-FORCE-SEND']);
 
@@ -539,7 +539,7 @@ class Plugin extends ServerPlugin {
 
                     if ($attendee->getNormalizedValue() === $message->recipient) {
                         if ($message->scheduleStatus) {
-                            $attendee['SCHEDULE-STATUS'] = $message->scheduleStatus;
+                            $attendee['SCHEDULE-STATUS'] = $message->getScheduleStatus();
                         }
                         unset($attendee['SCHEDULE-FORCE-SEND']);
                         break;