+++ tests/Sabre/DAV/Auth/Backend/AbstractDigestTest.php
@@ -116,7 +116,7 @@ class AbstractDigestTest extends \PHPUnit_Framework_TestCase {
 
         $backend = new AbstractDigestMock();
         $backend->setRealm('writing unittests on a saturday night');
-        $backend->requireAuth($request, $response);
+        $backend->challenge($request, $response);
 
         $this->assertStringStartsWith(
             'Digest realm="writing unittests on a saturday night"',